src/airuntime/requirements.txt
fastapi==0.116.2  
uvicorn[standard]==0.34.0
requests==2.32.4
pyjwt==2.9.0
google-adk==1.12.0
google-genai

